scraper/pyproject.toml
[project]
name = "chunimai-tracker"
version = "1.0.0"
description = "A project that track your playcount and rating on maimai and chunithm"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.30.0",
    "envparse>=0.2.0",
    "playwright>=1.52.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0.0",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-playwright>=0.7.2",
]
